Prune Opportunities exploration variants; keep the command center concept
The exploration produced many single-feature variants (worklist / battlefield /
sources / momentum) and stepping-stone pages (command-center v1+v2 / competitive
/ source-strategy). Now that the direction is settled, remove all of those and
their two multi-variant stories, keeping just the chosen action-first command
center (components/opportunities-playbook.tsx) behind one clean story
(Concepts/AEO command center). The live route components (opportunity-map,
opportunity-stability-chart) and the real-route QA stories are untouched.
